deleted the groebner_basis file and created a new file called gsolve that will eventually replace the functionality in groebner_class.py with a module of functions.
